+++ src/components/ActionIcon/ActionIcon.styles.ts
@@ -4,30 +4,92 @@ import { breakpointsPx } from "@/sizes";
 import { cssDimensions } from "@/utils/styles.utils";
 import { cssVarUsage } from "@utils";
 
-import type { ActionIconSize, ActionIconStyleProps } from "./ActionIcon.types";
-import type { SpacingLabel } from "@types";
+import type {
+  ActionIconColor,
+  ActionIconSize,
+  ActionIconStyleProps,
+} from "./ActionIcon.types";
+import type { ColorKey, PaletteColor, SpacingLabel } from "@types";
 import type { ValueMap } from "@ubloimmo/front-util";
 
 const sizeMap: ValueMap<ActionIconSize, SpacingLabel> = {
-  s: "s-4",
+  s: "s-6",
   m: "s-8",
   l: "s-10",
+};
+
+export const actionIconIconColorMap: ValueMap<
+  ActionIconSize,
+  Record<ActionIconColor, PaletteColor>
+> = {
+  s: {
+    white: "gray-600",
+    error: "error-base",
+    primary: "primary-base",
+  },
+  m: {
+    white: "gray-600",
+    error: "error-base",
+    primary: "primary-base",
+  },
+  l: {
+    white: "gray-800",
+    error: "gray-800",
+    primary: "primary-base",
+  },
+};
+
+const actionIconBackgroundMap: ValueMap<
+  ActionIconSize,
+  Record<ActionIconColor, PaletteColor | "white">
+> = {
+  s: {
+    white: "white",
+    error: "error-light",
+    primary: "primary-light",
+  },
+  m: {
+    white: "white",
+    error: "error-light",
+    primary: "primary-light",
+  },
+  l: {
+    white: "white",
+    error: "white",
+    primary: "primary-light",
+  },
+};
+
+const actionIconPaddingMap: ValueMap<ActionIconSize, SpacingLabel> = {
+  s: "s-05",
+  m: "s-1",
+  l: "s-2",
+};
+
+const actionIconBorderRadiusMap: ValueMap<ActionIconSize, SpacingLabel> = {
+  s: "s-05",
+  m: "s-1",
+  l: "s-1",
 };
 
 export const actionIconContainerStyles = ({
   $color,
   $size,
 }: ActionIconStyleProps) => {
-  const isSmall = $size === "s";
   const isMedium = $size === "m";
-  const lightColor = cssVarUsage(`${$color}-light`);
-  const borderColor = cssVarUsage(`${$color}-medium`);
-  const iconHoverColor = cssVarUsage(`${$color}-base`);
-  const borderColorTransparent = cssVarUsage(`${$color}-medium-00`);
-  const background = isSmall ? lightColor : "white";
+  const normalizedColor: ColorKey = $color === "white" ? "primary" : $color;
+  const iconHoverColor = cssVarUsage(`${normalizedColor}-base`);
+  const borderColorTransparent = cssVarUsage(`${normalizedColor}-medium-00`);
+  const borderColor =
+    $size === "l"
+      ? cssVarUsage(`${normalizedColor}-medium`)
+      : borderColorTransparent;
+  const backgroundColor = actionIconBackgroundMap[$size][$color];
+  const background =
+    backgroundColor === "white" ? "white" : cssVarUsage(backgroundColor);
 
-  const padding = cssVarUsage(`s-${isSmall ? "05" : isMedium ? "1" : "2"}`);
-  const borderRadius = cssVarUsage(`s-${isSmall ? "05" : "1"}`);
+  const padding = cssVarUsage(actionIconPaddingMap[$size]);
+  const borderRadius = cssVarUsage(actionIconBorderRadiusMap[$size]);
   const size = sizeMap[$size];
   const responsiveSize = isMedium ? sizeMap.l : size;
 
@@ -45,15 +107,19 @@ export const actionIconContainerStyles = ({
     transition: background 300ms ease-out 0s, border-color 300ms ease-out 0s,
       box-shadow 300ms ease-out 0s;
 
+    &:hover {
+      box-shadow: none;
+    }
+
     &:disabled {
       cursor: not-allowed;
+      box-shadow: none;
       background: var(--gray-50);
     }
 
     &:hover:not(:disabled) {
       transition-duration: 150ms;
       border-color: ${borderColor};
-      box-shadow: var(--shadow-card-elevation-medium);
 
       svg {
         fill: ${iconHoverColor};