+++ CHANGELOG.md
@@ -12,6 +12,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `v0.0.8` Form component
   - Do not render custom content if it itself doesn't (returns `null`)
 
+### Changed
+
+- `v0.0.4` ActionIcon component
+  - Update styles based on size and color according to figma update
+  - Update docs
+  - Fix size type declaration
+
+## 0.32.4 - 2024-09-19
+### Fixed
+- `v0.0.2` ComboBoxInput component
+  - Render passed html id
+
 ## 0.32.3 - 2024-09-18
 ### Changed
 - `v0.0.7` Form component